split the files as requested in review
